requirements.txt
# requirements.txt 
apache-airflow[postgres]~=2.8.4 
pandas~=2.0.0
numpy~=1.24.0
scikit-learn~=1.3.0 
psycopg2-binary
requests
sqlalchemy
pyarrow
xgboost
imblearn